Safari-611.1.10/LayoutTests/fast/dom/title-setter-new-text-node-expected.txt

Tests that setting document.title does not reuse title's text node child

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.title is "aaa"
PASS oldTextNode.textContent is "aaa"
document.title = 'bbb'
PASS oldTextNode.textContent is "aaa"
PASS oldTextNode is not newTextNode
PASS document.title is "bbb"
PASS newTextNode.textContent is "bbb"
PASS successfullyParsed is true

TEST COMPLETE
